Deleted changes in C and moved to B group
